t/tokenizer-test-charrefs.dat

#data
<!DOCTYPE html>&#0;
#errors
1:19:NULL;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFFFD"

#data
<!DOCTYPE html>&#x0;
#errors
1:19:NULL;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFFFD"

#data
<!DOCTYPE html>&#1;
#errors
1:19:invalid character reference:U+0001;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0001"

#data
<!DOCTYPE html>&#x1;
#errors
1:19:invalid character reference:U+0001;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0001"

#data
<!DOCTYPE html>&#2;
#errors
1:19:invalid character reference:U+0002;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0002"

#data
<!DOCTYPE html>&#x2;
#errors
1:19:invalid character reference:U+0002;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0002"

#data
<!DOCTYPE html>&#3;
#errors
1:19:invalid character reference:U+0003;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0003"

#data
<!DOCTYPE html>&#x3;
#errors
1:19:invalid character reference:U+0003;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0003"

#data
<!DOCTYPE html>&#4;
#errors
1:19:invalid character reference:U+0004;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0004"

#data
<!DOCTYPE html>&#x4;
#errors
1:19:invalid character reference:U+0004;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0004"

#data
<!DOCTYPE html>&#5;
#errors
1:19:invalid character reference:U+0005;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0005"

#data
<!DOCTYPE html>&#x5;
#errors
1:19:invalid character reference:U+0005;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0005"

#data
<!DOCTYPE html>&#6;
#errors
1:19:invalid character reference:U+0006;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0006"

#data
<!DOCTYPE html>&#x6;
#errors
1:19:invalid character reference:U+0006;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0006"

#data
<!DOCTYPE html>&#7;
#errors
1:19:invalid character reference:U+0007;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0007"

#data
<!DOCTYPE html>&#x7;
#errors
1:19:invalid character reference:U+0007;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0007"

#data
<!DOCTYPE html>&#8;
#errors
1:19:invalid character reference:U+0008;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0008"

#data
<!DOCTYPE html>&#x8;
#errors
1:19:invalid character reference:U+0008;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0008"

#data
<!DOCTYPE html>a&#9;
#errors
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "a\u0009"

#data
<!DOCTYPE html>a&#x9;
#errors
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "a\u0009"

#data
<!DOCTYPE html>a&#10;
#errors
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "a\u000A"

#data
<!DOCTYPE html>a&#xA;
#errors
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "a\u000A"

#data
<!DOCTYPE html>&#11;
#errors
1:19:invalid character reference:U+000B;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u000B"

#data
<!DOCTYPE html>&#xb;
#errors
1:19:invalid character reference:U+000B;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u000B"

#data
<!DOCTYPE html>a&#12;
#errors
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "a\u000C"

#data
<!DOCTYPE html>a&#xC;
#errors
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "a\u000C"

#data
<!DOCTYPE html>a&#13;
#errors
1:20:invalid character reference:U+000D;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "a\u000D"

#data
<!DOCTYPE html>a&#xd;
#errors
1:20:invalid character reference:U+000D;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "a\u000D"

#data
<!DOCTYPE html>a&#13;&#10;
#errors
1:20:invalid character reference:U+000D;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "a\u000D\u000A"

#data
<!DOCTYPE html>a&#xd;&#xa;
#errors
1:20:invalid character reference:U+000D;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "a\u000D\u000A"

#data
<!DOCTYPE html>&#14;
#errors
1:19:invalid character reference:U+000E;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u000E"

#data
<!DOCTYPE html>&#xe;
#errors
1:19:invalid character reference:U+000E;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u000E"

#data
<!DOCTYPE html>&#15;
#errors
1:19:invalid character reference:U+000F;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u000F"

#data
<!DOCTYPE html>&#xf;
#errors
1:19:invalid character reference:U+000F;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u000F"

#data
<!DOCTYPE html>&#16;
#errors
1:19:invalid character reference:U+0010;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0010"

#data
<!DOCTYPE html>&#x10;
#errors
1:19:invalid character reference:U+0010;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0010"

#data
<!DOCTYPE html>&#17;
#errors
1:19:invalid character reference:U+0011;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0011"

#data
<!DOCTYPE html>&#x11;
#errors
1:19:invalid character reference:U+0011;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0011"

#data
<!DOCTYPE html>&#18;
#errors
1:19:invalid character reference:U+0012;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0012"

#data
<!DOCTYPE html>&#x12;
#errors
1:19:invalid character reference:U+0012;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0012"

#data
<!DOCTYPE html>&#19;
#errors
1:19:invalid character reference:U+0013;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0013"

#data
<!DOCTYPE html>&#x13;
#errors
1:19:invalid character reference:U+0013;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0013"

#data
<!DOCTYPE html>&#20;
#errors
1:19:invalid character reference:U+0014;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0014"

#data
<!DOCTYPE html>&#x14;
#errors
1:19:invalid character reference:U+0014;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0014"

#data
<!DOCTYPE html>&#21;
#errors
1:19:invalid character reference:U+0015;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0015"

#data
<!DOCTYPE html>&#x15;
#errors
1:19:invalid character reference:U+0015;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0015"

#data
<!DOCTYPE html>&#22;
#errors
1:19:invalid character reference:U+0016;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0016"

#data
<!DOCTYPE html>&#x16;
#errors
1:19:invalid character reference:U+0016;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0016"

#data
<!DOCTYPE html>&#23;
#errors
1:19:invalid character reference:U+0017;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0017"

#data
<!DOCTYPE html>&#x17;
#errors
1:19:invalid character reference:U+0017;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0017"

#data
<!DOCTYPE html>&#24;
#errors
1:19:invalid character reference:U+0018;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0018"

#data
<!DOCTYPE html>&#x18;
#errors
1:19:invalid character reference:U+0018;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0018"

#data
<!DOCTYPE html>&#25;
#errors
1:19:invalid character reference:U+0019;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0019"

#data
<!DOCTYPE html>&#x19;
#errors
1:19:invalid character reference:U+0019;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0019"

#data
<!DOCTYPE html>&#26;
#errors
1:19:invalid character reference:U+001A;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u001A"

#data
<!DOCTYPE html>&#x1A;
#errors
1:19:invalid character reference:U+001A;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u001A"

#data
<!DOCTYPE html>&#27;
#errors
1:19:invalid character reference:U+001B;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u001B"

#data
<!DOCTYPE html>&#x1B;
#errors
1:19:invalid character reference:U+001B;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u001B"

#data
<!DOCTYPE html>&#28;
#errors
1:19:invalid character reference:U+001C;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u001C"

#data
<!DOCTYPE html>&#x1C;
#errors
1:19:invalid character reference:U+001C;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u001C"

#data
<!DOCTYPE html>&#29;
#errors
1:19:invalid character reference:U+001D;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u001D"

#data
<!DOCTYPE html>&#x1D;
#errors
1:19:invalid character reference:U+001D;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u001D"

#data
<!DOCTYPE html>&#30;
#errors
1:19:invalid character reference:U+001E;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u001E"

#data
<!DOCTYPE html>&#x1E;
#errors
1:19:invalid character reference:U+001E;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u001E"

#data
<!DOCTYPE html>&#31;
#errors
1:19:invalid character reference:U+001F;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u001F"

#data
<!DOCTYPE html>&#x1F;
#errors
1:19:invalid character reference:U+001F;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u001F"

#data
<!DOCTYPE html>&#127;
#errors
1:19:invalid character reference:U+007F;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u007F"

#data
<!DOCTYPE html>&#x7F;
#errors
1:19:invalid character reference:U+007F;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u007F"

#data
<!DOCTYPE html>&#128;
#errors
1:19:invalid character reference:U+0080;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u20AC"

#data
<!DOCTYPE html>&#x80;
#errors
1:19:invalid character reference:U+0080;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u20AC"

#data
<!DOCTYPE html>&#129;
#errors
1:19:invalid character reference:U+0081;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0081"

#data
<!DOCTYPE html>&#x81;
#errors
1:19:invalid character reference:U+0081;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0081"

#data
<!DOCTYPE html>&#130;
#errors
1:19:invalid character reference:U+0082;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u201A"

#data
<!DOCTYPE html>&#x82;
#errors
1:19:invalid character reference:U+0082;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u201A"

#data
<!DOCTYPE html>&#131;
#errors
1:19:invalid character reference:U+0083;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0192"

#data
<!DOCTYPE html>&#x83;
#errors
1:19:invalid character reference:U+0083;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0192"

#data
<!DOCTYPE html>&#131;
#errors
1:19:invalid character reference:U+0083;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0192"

#data
<!DOCTYPE html>&#x83;
#errors
1:19:invalid character reference:U+0083;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0192"

#data
<!DOCTYPE html>&#133;
#errors
1:19:invalid character reference:U+0085;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2026"

#data
<!DOCTYPE html>&#x85;
#errors
1:19:invalid character reference:U+0085;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2026"

#data
<!DOCTYPE html>&#134;
#errors
1:19:invalid character reference:U+0086;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2020"

#data
<!DOCTYPE html>&#x86;
#errors
1:19:invalid character reference:U+0086;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2020"

#data
<!DOCTYPE html>&#135;
#errors
1:19:invalid character reference:U+0087;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2021"

#data
<!DOCTYPE html>&#x87;
#errors
1:19:invalid character reference:U+0087;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2021"

#data
<!DOCTYPE html>&#136;
#errors
1:19:invalid character reference:U+0088;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u02C6"

#data
<!DOCTYPE html>&#x88;
#errors
1:19:invalid character reference:U+0088;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u02C6"

#data
<!DOCTYPE html>&#137;
#errors
1:19:invalid character reference:U+0089;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2030"

#data
<!DOCTYPE html>&#x89;
#errors
1:19:invalid character reference:U+0089;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2030"

#data
<!DOCTYPE html>&#138;
#errors
1:19:invalid character reference:U+008A;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0160"

#data
<!DOCTYPE html>&#x8A;
#errors
1:19:invalid character reference:U+008A;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0160"

#data
<!DOCTYPE html>&#139;
#errors
1:19:invalid character reference:U+008B;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2039"

#data
<!DOCTYPE html>&#x8B;
#errors
1:19:invalid character reference:U+008B;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2039"

#data
<!DOCTYPE html>&#140;
#errors
1:19:invalid character reference:U+008C;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0152"

#data
<!DOCTYPE html>&#x8C;
#errors
1:19:invalid character reference:U+008C;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0152"

#data
<!DOCTYPE html>&#141;
#errors
1:19:invalid character reference:U+008D;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u008D"

#data
<!DOCTYPE html>&#x8D;
#errors
1:19:invalid character reference:U+008D;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u008D"

#data
<!DOCTYPE html>&#142;
#errors
1:19:invalid character reference:U+008E;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u017D"

#data
<!DOCTYPE html>&#x8E;
#errors
1:19:invalid character reference:U+008E;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u017D"

#data
<!DOCTYPE html>&#143;
#errors
1:19:invalid character reference:U+008F;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u008F"

#data
<!DOCTYPE html>&#x8F;
#errors
1:19:invalid character reference:U+008F;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u008F"

#data
<!DOCTYPE html>&#144;
#errors
1:19:invalid character reference:U+0090;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0090"

#data
<!DOCTYPE html>&#x90;
#errors
1:19:invalid character reference:U+0090;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0090"

#data
<!DOCTYPE html>&#145;
#errors
1:19:invalid character reference:U+0091;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2018"

#data
<!DOCTYPE html>&#x91;
#errors
1:19:invalid character reference:U+0091;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2018"

#data
<!DOCTYPE html>&#146;
#errors
1:19:invalid character reference:U+0092;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2019"

#data
<!DOCTYPE html>&#x92;
#errors
1:19:invalid character reference:U+0092;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2019"

#data
<!DOCTYPE html>&#147;
#errors
1:19:invalid character reference:U+0093;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u201C"

#data
<!DOCTYPE html>&#x93;
#errors
1:19:invalid character reference:U+0093;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u201C"

#data
<!DOCTYPE html>&#148;
#errors
1:19:invalid character reference:U+0094;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u201D"

#data
<!DOCTYPE html>&#x94;
#errors
1:19:invalid character reference:U+0094;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u201D"

#data
<!DOCTYPE html>&#149;
#errors
1:19:invalid character reference:U+0095;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2022"

#data
<!DOCTYPE html>&#x95;
#errors
1:19:invalid character reference:U+0095;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2022"

#data
<!DOCTYPE html>&#150;
#errors
1:19:invalid character reference:U+0096;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2013"

#data
<!DOCTYPE html>&#x96;
#errors
1:19:invalid character reference:U+0096;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2013"

#data
<!DOCTYPE html>&#151;
#errors
1:19:invalid character reference:U+0097;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2014"

#data
<!DOCTYPE html>&#x97;
#errors
1:19:invalid character reference:U+0097;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2014"

#data
<!DOCTYPE html>&#152;
#errors
1:19:invalid character reference:U+0098;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u02DC"

#data
<!DOCTYPE html>&#x98;
#errors
1:19:invalid character reference:U+0098;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u02DC"

#data
<!DOCTYPE html>&#153;
#errors
1:19:invalid character reference:U+0099;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2122"

#data
<!DOCTYPE html>&#x99;
#errors
1:19:invalid character reference:U+0099;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u2122"

#data
<!DOCTYPE html>&#154;
#errors
1:19:invalid character reference:U+009A;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0161"

#data
<!DOCTYPE html>&#x9A;
#errors
1:19:invalid character reference:U+009A;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0161"

#data
<!DOCTYPE html>&#155;
#errors
1:19:invalid character reference:U+009B;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u203A"

#data
<!DOCTYPE html>&#x9B;
#errors
1:19:invalid character reference:U+009B;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u203A"

#data
<!DOCTYPE html>&#156;
#errors
1:19:invalid character reference:U+009C;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0153"

#data
<!DOCTYPE html>&#x9C;
#errors
1:19:invalid character reference:U+009C;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0153"

#data
<!DOCTYPE html>&#157;
#errors
1:19:invalid character reference:U+009D;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u009D"

#data
<!DOCTYPE html>&#x9D;
#errors
1:19:invalid character reference:U+009D;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u009D"

#data
<!DOCTYPE html>&#158;
#errors
1:19:invalid character reference:U+009E;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u017E"

#data
<!DOCTYPE html>&#x9E;
#errors
1:19:invalid character reference:U+009E;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u017E"

#data
<!DOCTYPE html>&#159;
#errors
1:19:invalid character reference:U+009F;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0178"

#data
<!DOCTYPE html>&#x9F;
#errors
1:19:invalid character reference:U+009F;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\u0178"

#data
<!DOCTYPE html>&#55361;
#errors
1:19:invalid character reference:U+D841;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFFFD"

#data
<!DOCTYPE html>&#xD841;
#errors
1:19:invalid character reference:U+D841;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFFFD"

#data
<!DOCTYPE html>&#64996;
#errors
1:19:invalid character reference:U+FDE4;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFDE4"

#data
<!DOCTYPE html>&#xFDE4;
#errors
1:19:invalid character reference:U+FDE4;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFDE4"

#data
<!DOCTYPE html>&#65533;
#errors
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFFFD"

#data
<!DOCTYPE html>&#xFFFD;
#errors
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFFFD"

#data
<!DOCTYPE html>&#65534;
#errors
1:19:invalid character reference:U+FFFE;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFFFE"

#data
<!DOCTYPE html>&#xFFFE;
#errors
1:19:invalid character reference:U+FFFE;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFFFE"

#data
<!DOCTYPE html>&#65535;
#errors
1:19:invalid character reference:U+FFFF;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFFFF"

#data
<!DOCTYPE html>&#xFFFF;
#errors
1:19:invalid character reference:U+FFFF;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFFFF"

#data
<!DOCTYPE html>&#1114110;
#errors
1:19:invalid character reference:U+10FFFE;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\U0010FFFE"

#data
<!DOCTYPE html>&#x10FFFE;
#errors
1:19:invalid character reference:U+10FFFE;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\U0010FFFE"

#data
<!DOCTYPE html>&#1114111;
#errors
1:19:invalid character reference:U+10FFFF;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\U0010FFFF"

#data
<!DOCTYPE html>&#x10FFFF;
#errors
1:19:invalid character reference:U+10FFFF;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\U0010FFFF"

#data
<!DOCTYPE html>&#1114112;
#errors
1:19:invalid character reference:U-00110000;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFFFD"

#data
<!DOCTYPE html>&#x110000;
#errors
1:19:invalid character reference:U-00110000;m
#document escaped
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "\uFFFD"
